+++ Verve/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.18.1.1
+* Add support for binary consent state as a fallback option if the TCFv2 GDPR consent string is not present in Shared Preferences.
+* Remove unnecessary MAX SDK version check.
+
 ## 2.18.1.0
 * Certified with Verve SDK 2.18.1.